CaDrA Search

Performs a permutation-based testing on sample permutation of observed input scores using candidate_search() as the main iterative function for each run.

CaDrA(
  ES,
  input_score,
  method = c("ks", "wilcox", "revealer", "custom"),
  custom_function = NULL,
  custom_parameters = NULL,
  alternative = c("less", "greater", "two.sided"),
  metric = c("pval", "stat"),
  weights = NULL,
  top_N = 1,
  search_start = NULL,
  search_method = c("both", "forward"),
  max_size = 7,
  n_perm = 1000,
  smooth = TRUE,
  obs_best_score = NULL,
  plot = TRUE,
  ncores = 1,
  cache_path = NULL,
  verbose = FALSE
)

Arguments

ES

an expression set of binary features (required). It must be a BioBase expressionSet object. The rownames of the expression set must contain unique features which are used to search for best features.

input_score

a vector of continuous values of a response of interest (required). The input_score object must have names or labels that match the colnames of the expression matrix.

method

a character string specifies a method to compute the score for each feature ("ks" or "wilcox" or "revealer" (conditional mutual information from REVEALER) or "custom" (a customized scoring method)). Default is ks.

custom_function

if method is "custom", specifies the customized function here. Default is NULL.

custom_parameters

if method is "custom", specifies a list of arguments to be passed to the custom_function(). Default is NULL.

alternative

a character string specifies an alternative hypothesis testing ("two.sided" or "greater" or "less"). Default is less for left-skewed significance testing.

metric

a character string specifies a metric to search for best features. "pval" or "stat" may be used which corresponding to p-value or score statistic. Default is pval. Note: Revealer method only utilized score statistics values (no p-value).

weights

if method is ks, specifies a vector of weights will perform a weighted-KS testing. Default is NULL.

top_N

if search_start is NULL, then top_N can be used here. top_N can be an integer specifies the number of features to start the search over. Starting from the top to 'N' features. Default is NULL.

search_start

if top_N is NULL, then search_start can be used here. search_start can be an integer specifies an index within the expression set object of which feature to start the search with. Default is NULL. If NULL, then the search starts with the top ranked feature. If an integer is specified (N, where N < nrow(ES)), the search starts with the Nth best feature. If a string is specified, the search starts with the feature with this name (must be a valid rowname in ES)

search_method

a character string specifies an algorithm to filter out the best candidates ("forward" or "both"). Default is both (i.e., backward and forward).

max_size

an integer specifies a maximum size that a meta-feature can extend to do for a given search. Default is 7.

n_perm

an integer specifies the number of permutations to perform. Default is 1000.

smooth

a logical value indicates whether or not to smooth the p-value calculation to avoid p-value of 0. Default is TRUE.

obs_best_score

a numeric value corresponding to the observed (best) candidate search score to use for permutation based p-value computation. Default is NULL. If set to NULL, we will compute the observed best score based on the given input_score and ES variables.

plot

a logical value indicates whether or not to plot the empirical null distribution with the observed score and permutation p-value. Default is TRUE.

ncores

an integer specifies the number of cores to perform parallelization for permutation testing. Default is 1.

cache_path

a full path uses to cache permutation-based score distributions. If the permutation for a given ES and its dependent search variables such as 'top_N' exist, we recycle these values instead of re-computing them to save time. Default is NULL. If NULL, the cache path is set to ~/.Rcache for future loading.

verbose

a logical value indicates whether or not to print the diagnostic messages. Default is FALSE.

Value

a list of key parameters that were used to cache the permutation test, permutation best scores for n_perm, a permutation p-value, and observed best score

Author

Reina Chau

Examples

# \donttest{
# Load R library
library(Biobase)
#> Loading required package: BiocGenerics
#> Warning: package ‘BiocGenerics’ was built under R version 4.0.5
#> Loading required package: parallel
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

# Load pre-computed expression set
data(sim.ES)

# Set seed
set.seed(123)

# Provide a vector of continuous scores for a target profile with names 
# to each score value 
input_score = rnorm(n = ncol(sim.ES))
names(input_score) <- colnames(sim.ES)

# Define additional parameters and start the function
# This function takes some time to run
 cadra_result <- CaDrA(
  ES = sim.ES, input_score = input_score, method = "ks", weights = NULL,
  alternative = "less", metric = "pval", top_N = 1, 
  search_start = NULL, search_method = "both", max_size = 7, n_perm = 1000, 
  plot = TRUE, smooth = TRUE, obs_best_score = NULL, 
  ncores = 1, cache_path = NULL
)
#> Setting cache root path as: /Users/reinachau/Library/Caches/org.R-project.R/R/R.cache
#> Found 1000 cached permutation-based scores for the specified dataset and search parameters..
#> Loading permutation scores from cache..
#> FINISHED
#> Time elapsed: 0 mins 
#> 
#> Computing observed best score ..
#> 
#> Number of permutation-based scores being considered: 1000
#> Permutation p-value: 0.955044955044955
#> 
# }