CaDrA Search

Performs a permutation-based testing using results from candidate_search() function.

cadra_search(
  ES,
  input_score,
  method = "ks",
  custom_function = NULL,
  custom_parameters = NULL,
  alternative = "less",
  metric = "pval",
  weights = NULL,
  target_match = "positive",
  top_N = NULL,
  search_start = NULL,
  search_method = "both",
  max_size = 7,
  n_perm = 1000,
  seed = 123,
  smooth = TRUE,
  obs_best_score = NULL,
  plot = TRUE,
  ncores = 1,
  cache_path = NULL,
  return_perm_pval = TRUE
)

Arguments

ES

an expression set of binary features (required). It must be a BioBase expressionSet object. The rownames of the expression set must contain unique features which are used to search for best features.

input_score

a vector of continuous values for a target profile (required). The input_score object must have names or labels that match the colnames of the expression matrix.

method

a character string specifies a method to compute the score for each feature ("ks" or "wilcox" or "revealer" (conditional mutual information from REVEALER) or "custom" (a customized method)). Default is ks.

custom_function

if method is "custom", specifies the customized function here. Default is NULL.

custom_parameters

if method is "custom", specifies a list of arguments to be passed to the custom_function(). Default is NULL.

alternative

a character string specifies an alternative hypothesis testing ("two.sided" or "greater" or "less"). Default is less for left-skewed significance testing.

metric

a character string specifies a metric to use for candidate search criteria. "pval" or "stat" may be used, corresponding to the score p-value or statistic. Default is pval.

weights

a vector of weights use to perform a weighted-KS testing. Default is NULL.

target_match

a direction of target matching ("negative" or "positive") from REVEALER. Use "positive" to match the higher values of the target, "negative" to match the lower values. Default is positive.

top_N

an integer specifies the number of features to start the search over, starting from the top 'N' features in each case. Default is NULL.

search_start

an integer specifies an index within the expression set object of which feature to start the candidate search with. Default is NULL. If NULL, then the search starts with the top ranked feature. If an integer is specified (N, where N < nrow(ES)), the search starts with the Nth best feature. If a string is specified, the search starts with the feature with this name (must be a valid rowname in ES)

search_method

a character string specifies a method to filter out the best candidates ("forward" or "both"). Default is both (backward and forward).

max_size

an integer specifies a maximum size that a meta-feature can extend to do for a given search. Default is 7.

n_perm

an integer specifies the number of permutations to perform. Default is 1000.

seed

a seed set for permutation. Default is 123.

smooth

a logical value indicates whether or not to smooth the p-value calculation to avoid p-value of 0. Default is TRUE.

obs_best_score

a numeric value corresponding to the observed (best) candidate search score to use for permutation based p-value computation. Default is NULL. If set to NULL, we compute the observed score given the input_score and ES variables.

plot

a logical value indicates whether or not to plot the empirical null distribution with the observed score and permutation p-value. Default is TRUE.

ncores

an integer specifies the number of cores to perform parallelization for permutation testing. Default is 1.

cache_path

a full path uses to cache permutation-based score distributions. If the permutation for a given ES and its dependent search variables such as 'N' exist, we recycle these values instead of re-computing them to save time. Default is NULL. If NULL, the cache path is set to ~/.Rcache for future loading.

return_perm_pval

a logical value indicates whether or not to return the permutation-based p-value computed by the function. Default is TRUE.

Value

If return_perm_pval is set to TRUE, it will return a permutation p-value.

Author

Reina Chau

Examples

# Load R library
library(Biobase)
#> Loading required package: BiocGenerics
#> Loading required package: parallel
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

# Load pre-computed expression set
data(sim.ES)

# set seed
set.seed(123)

# Provide a vector of continuous scores for a target profile with names to each score value 
input_score = rnorm(n = ncol(sim.ES))
names(input_score) <- colnames(sim.ES)

# Define additional parameters and start the function
# Not run as this would take some time to run
# candidate_search_result <- cadra_search(
#  ES = sim.ES, input_score = input_score, method = "ks", weights = NULL,
#  alternative = "less", metric = "pval", top_N = NULL, 
#  search_start = NULL, search_method = "both", max_size = 7, n_perm = 1000, 
#  seed = 123, plot = TRUE, smooth = TRUE, obs_best_score = NULL, 
#  ncores = 1, cache_path = NULL, return_perm_pval = TRUE
#)