Candidate Search

Performs heuristic search on a set of binary features to determine whether there are features whose union is more skewed (enriched at the extremes) than either features alone. This is the main functionality of the CaDrA package.

candidate_search(
  ES,
  input_score,
  method = "ks",
  custom_function = NULL,
  custom_parameters = NULL,
  alternative = "less",
  metric = "stat",
  weights = NULL,
  search_start = NULL,
  top_N = 1,
  search_method = "both",
  max_size = 7,
  best_score_only = FALSE,
  do_plot = TRUE,
  verbose = FALSE
)

Arguments

ES

an expression set of binary features (required). It must be a BioBase ExpressionSet object. The rownames of the expression set must contain unique features which are used in the search.

input_score

a vector of continuous scores of a functional response of interest (required). The input_score must have names or labels that matches the colnames of the expression matrix.

method

a character string specifies a the scoring function that is used in the search. There are 4 options: ("ks" or "wilcox" or "revealer" (conditional mutual information from REVEALER) or "custom" (a customized scoring method)). Default is ks.

custom_function

if method is "custom", specifies the customized function here. Default is NULL.

custom_parameters

if method is "custom", specifies a list of arguments to be passed to the custom_function(). Default is NULL.

alternative

a character string specifies an alternative hypothesis testing ("two.sided" or "greater" or "less"). Default is less for left-skewed significance testing.

metric

a character string specifies a metric to search for best features. "pval" or "stat" may be used which corresponding to p-value or score statistic. Default is pval. Note: Revealer method only utilized score statistics values (no p-value).

weights

if method is ks, specifies a vector of weights will perform a weighted-KS testing. Default is NULL.

search_start

an integer specifies an index within the expression set object of which feature(s) to start the candidate search with. Default is NULL. If NULL, then the search starts with the top ranked feature. If an integer is specified (N, where N < nrow(ES)), the search starts with the Nth best feature. If a string is specified, the search starts with the feature with this name (must be a valid row name in ES)

top_N

an integer specifies the number of features to start the search over, starting from the top 'N' features in each case. Default is 1.

search_method

a character string specifies an algorithm to filter out the best features ("forward" or "both"). Default is both (i.e. backward and forward).

max_size

an integer specifies a maximum size that a meta-feature can extend to do for a given search. Default is 7.

best_score_only

a logical value indicates whether or not the function should return only the score corresponding to the search results. Default is FALSE.

do_plot

a logical value indicates whether or not to plot the resulting evaluation matrix. Default is TRUE.

verbose

a logical value indicates whether or not to print the diagnostic messages. Default is FALSE.

Value

If best_score_only is set to TRUE, the function returns a list object with the score corresponding to the union of the search meta-feature. If best_score_only is set to FALSE, a list containing the ES object (pertaining to the returned meta-feature) as well as its corresponding score and observed input scores are returned.

Examples


# Load R library
library(Biobase)

# Load pre-computed expression set
data(sim.ES)

# set seed
set.seed(123)

# Provide a vector of continuous scores for a target profile 
# The scores must have labels or names that match the colnames of expression matrix
input_score = rnorm(n = ncol(sim.ES))
names(input_score) <- colnames(sim.ES)

# Define additional parameters and run the function
candidate_search_result <- candidate_search(
  ES = sim.ES, input_score = input_score, method = "ks", 
  alternative = "less", weights = NULL, metric = "pval", 
  search_start = NULL, top_N = 1, search_method = "both", 
  max_size = 7, best_score_only = FALSE
)