Compute a Kolmogorov-Smirnov score for a given ranked list

Compute a Kolmogorov-Smirnov score for a given ranked list

ks_gene_score(
  n.x,
  y,
  weights = NULL,
  weight_p = 1,
  alternative = "less",
  do_pval = TRUE,
  absolute = FALSE,
  exact = NULL,
  plot_dat = FALSE
)

Arguments

n.x

length of a ranked list

y

positions of interested geneset items in the ranked list

weights

weights for weighted score (see Subramanian et al.) (usually, sort(score))

weight_p

weights exponent

alternative

alternative hypothesis for p-value calculation

do_pval

compute asymptotic p-value

absolute

takes max - min score rather than the maximum deviation from null

exact

compute exact p-value

plot_dat

return dataframe of x and y axis data for ES plot

Value

a data frame with two columns: score and p_value