ks.genescore wrapper

Compute directional Kolmogorov-Smirnov scores for each row of a given vector

Compute directional Kolmogorov-Smirnov scores for each row of a given vector

ks_genescore_wrap(n_x, y, weight, alt = "less")

ks_genescore_wrap(n_x, y, weight, alt = "less")

Arguments

n_x

length of ranked list

y

positions of geneset items in ranked list (ranks)

weight

a vector of weights

alt

alternative hypothesis for p-value calculation