Pre-filter features

Pre-filter a dataset prior to running step-wise heuristic search in order to avoid testing features that are too prevalent or too sparse across samples in the dataset

Pre-filter a dataset prior to running step-wise heuristic search in order to avoid testing features that are too prevalent or too sparse across samples in the dataset

prefilter_data(ES, max.cutoff = 0.6, min.cutoff = 0.03)

prefilter_data(ES, max.cutoff = 0.6, min.cutoff = 0.03)

Arguments

ES

an expression set object containing binary features used for step-wise search

max.cutoff

a numeric value between 0 and 1 describing the absolute prevalence of a feature across all samples in the dataset above which the feature will be filtered out. Default is 0.6 (feature that occur in 60 percent or more of the samples will be removed)

min.cutoff

a numeric value between 0 and 1 describing the absolute prevalence of a feature across all samples in the dataset below which the feature will be filtered out. Default is 0.03 (feature that occur in 3 percent or less of the samples will be removed)

Value

An expression set object with only the filtered-in features given the filter thresholds specified An expression set object with only the filtered-in features given the filter thresholds specified

Examples

data(sim.ES)

# Filter out features having < 3 and > 60% prevalence across all samples (default)
sim.ES.filt1 <- prefilter_data(sim.ES)
#> Pre-filtering features ..
#> 
#> Removing features having <  3 and >  60  % occurence in sample set..
#> 883  features retained out of  1000  supplied features in dataset
#> 

# Change the min cut-off to 1% prevalence, instead of the default 3%
sim.ES.filt2 <- prefilter_data(sim.ES,min.cutoff=0.01)
#> Pre-filtering features ..
#> 
#> Removing features having <  1 and >  60  % occurence in sample set..
#> 1000  features retained out of  1000  supplied features in dataset
#> 

# Change the max cut-off to 65% prevalence, instead of the default 60%
sim.ES.filt3 <- prefilter_data(sim.ES,max.cutoff=0.65) 
#> Pre-filtering features ..
#> 
#> Removing features having <  3 and >  65  % occurence in sample set..
#> 883  features retained out of  1000  supplied features in dataset
#> 

data(sim.ES)

# Filter out features having < 3 and > 60% prevalence across all samples (default)
sim.ES.filt1 <- prefilter_data(sim.ES)
#> Pre-filtering features ..
#> 
#> Removing features having <  3 and >  60  % occurence in sample set..
#> 883  features retained out of  1000  supplied features in dataset
#> 

# Change the min cut-off to 1% prevalence, instead of the default 3%
sim.ES.filt2 <- prefilter_data(sim.ES,min.cutoff=0.01)
#> Pre-filtering features ..
#> 
#> Removing features having <  1 and >  60  % occurence in sample set..
#> 1000  features retained out of  1000  supplied features in dataset
#> 

# Change the max cut-off to 65% prevalence, instead of the default 60%
sim.ES.filt3 <- prefilter_data(sim.ES,max.cutoff=0.65) 
#> Pre-filtering features ..
#> 
#> Removing features having <  3 and >  65  % occurence in sample set..
#> 883  features retained out of  1000  supplied features in dataset
#>