Step-wise search

Performs step-wise heuristic search using an ordered set of binary features to see whether there are features whose union is more skewed (enriched at the extremes) than either features alone. This is the main functionality of the CaDrA package.

stepwise.search(
  ranking = NULL,
  ES,
  max.size = 7,
  metric = "pval",
  method = c("ks", "wilcox"),
  back_search = TRUE,
  cust_start = NULL,
  best.score.only = FALSE,
  alt = "less",
  wts = NULL,
  rnks = NULL,
  verb = FALSE
)

Arguments

ranking

vector containing rankings for sample ordering. Default is NULL. If NULL, we assume the ESet is already ranked

ES

an expression set object of binary features. The first column of the featureData for the expression set must contain the names of the corresponding features, which are used in the search

max.size

an integer specifying the maximum size a meta-feature can extend do for a given search. Default is 7

metric

a character string specifying which metric to use for stepwise search criteria. One of either 'pval' or 'stat' may be used, corresponding to the score p-value or statistic. Default is 'pval'

method

a character string specifying the method used to compute scores for features, must be one of "ks" or "wilcox"

back_search

a logical indicating whether or not to perform a forward-backward search (i.e. remove features along the search if it improves the best score). Default is TRUE. Uses function backward_check()

cust_start

an integer specifying a specific index within the expression set object of the feature to start the step-wise search with. Default is NULL. If NULL, then the search starts with the top ranked feature. If an integer is specified (N, where N < nrow(dataset)), then the search starts with the Nth best feature. If a string is specified, then the search starts with the feature with this name (must be a valid rowname in the dataset)

best.score.only

a logical indicating whether or not the function should return only the score corresponding to the search results. Default is FALSE

alt

a character string specifying the alternative hypothesis, must be one of "two.sided","greater" or "less". Default is "less" for left-skewed significance testing.

wts

an integer vector of weights to use if performing weighted-KS testing. Default is NULL. Value passed to compute_score() function

rnks

an integer vector of sample rankings to use if performing Wilcoxon rank sum testing. Default is NULL. If NULL, then samples are assumed to be ordered by increasing ranking. Value passed to compute_score() function

verb

a logical indicating whether or not to print diagnostic messages. Default is FALSE

Value

If best.score.only is set to TRUE, this function returns a list object with the score corresponding to the union of the search meta-feature. If this is set to FALSE, a list containing both the ESet object pertaining to the returned meta-feature as well as the corresponding score is returned.