Wilcoxon Rank Sum Scoring Method

Compute directional Wilcoxon rank sum score for each row of a given binary matrix

wilcox_genescore_mat(
  mat,
  ranks = NULL,
  alternative = c("two.sided", "greater", "less"),
  verbose = FALSE
)

Arguments

mat

a matrix of binary features (required).

ranks

a vector of sample rankings use to perform the Wilcoxon test. Default is NULL. If NULL, then samples are assumed to be ordered by increasing rankings. If not NULL, ranks must include labels or names that associated with the colnames of the feature matrix.

alternative

a character string specifies an alternative hypothesis testing ("two.sided" or "greater" or "less"). Default is less for left-skewed significance testing.

verbose

a logical value indicates whether or not to print the diagnostic messages. Default is FALSE.

Value

A data frame with two columns: score and p_value

Examples

# Load R library
library(Biobase)
# Load pre-computed expression set
data(sim.ES)

# Define additional parameters and run the function
wilcox_genescore_mat_result <- wilcox_genescore_mat(
  mat = exprs(sim.ES), 
  ranks = NULL,
  alternative = "less"
)